chore: refactor to monorepo
