+++ README.md
@@ -9,31 +9,34 @@ This is created as a learning opportunity to:
 1. improve Dockerfile / docker compose knowledge
 1. ultimately have a fun playground / home integration tool on a server somewhere
 
-## Prerequisites
+## Local development/contribution
+
+### Prerequisites
 
 - Docker (and docker-compose)
 - Deno v1.1.1 and above
 - VSCode will prompt you to install specific extensions, it is advised to at least install the Deno extension
 
-## Local development/contribution
-
 _NOTE: further instructions are written towards UNIX systems since .sh scripts are used. But this probably all works on Windows as well._
 
 ### Testing inside Docker
 
-- `./buildAndRun.sh`
+`./denofn.sh` takes various commands, depending on what you want to do:
 
-### Testing on you machine
-
-- `mkdir registry && mkdir registry_in`
-- place your functions in `registry_in`
-- `./registry.sh`
-- `./execution.sh`
+- `build` uses docker-compose to build all docker images
+- `up <PATH>` creates volumes at selected path and docker-compose up
+- `down` docker-compose down + removes volumes
+- `update` down + build + up
+- `log` docker-compose logs
+- `clear <PATH>` clears registry folder at selected path
+- `cache <reload|update>`
+  - `reload` will fetch all dependencies on your local machine, using the lock.json. Kind of like `npm install/yarn`
+  - `update` will fetch updates and new dependencies and add them to lock.json. Kind of like `npm install <packages>/yarn add <packages>`
 
 ### Contributing new deps
 
 - Non versioned dependencies will not be allowed
-- run `./cache.sh` for opening a PR
+- run `./denofn.sh cache update` for opening a PR
 
 ## Function layout
 
@@ -63,14 +66,12 @@ TODO: use deno.land or github/unpkg in example to micro/mod.ts
 
 ```json
 {
-  "permissions": [],
+  "whitelist": [],
   "warmupOnStart": true
 }
 ```
 
-TODO: Permission provision in registry
-
-- `permissions`: `deno run` permissions. Will probably disallow `--allow-run`
+- `whitelist`: all domains that are allowed for `--allow-net`
 - `warmupOnStart`: will bootstrap your function immediately upon docker initialization. This is still a bit flaky as there are race conditions with registration, will be fixed soon.
 
 ## Todos
@@ -78,12 +79,8 @@ TODO: Permission provision in registry
 1. [ ] Functions Registry
    1. [x] Bundle with Deno
    1. [x] Port provision in registry
-   1. [ ] Permission provision in registry
+   1. [x] Whitelist provision in registry
    1. [ ] Allow serving static files (map asset name to filename?)
-   1. [ ] Handle functions without --allow-net?
-      - Assume they run in one go. No warmup needed. Pure process spawning, calculation, and done.
-      - Different type of handler needed as well
-      - Override console to combat fiddling with output
 1. [ ] Spawn, execute, kill
    1. [x] Execution server
    1. [x] Spawn subprocess for requests
@@ -99,13 +96,16 @@ TODO: Permission provision in registry
       1. [ ] Remove registration via mounted volume
       1. [ ] Refresh code
    1. [ ] Monitoring dashboard UI
-1. [x] Docker
-   1. [x] Create Dockerfile for registry
-   1. [x] Create Dockerfile for execution
-   1. [x] Create docker-compose file
+1. [ ] extract orchestration layer to separate module
+   1. [ ] try toAsyncIterator for spawning deno processes
+   1. [ ] spawning and killing shouldn't exist "inside" execution layer
+   1. [ ] potentially extract into separate dockerfile as well
 1. [ ] Lib improvements
    1. [ ] Opine is quite messy in implementation (see registerScriptHandler.ts)
+   1. [ ] Publish script for packages (versioned release)
+      - All packages on tag denofn@1.0.0 will have its sibling packages as denofn@1.0.0 in their respective deps.ts file
+      - When performing new development, the same script should reset these dependencies to @main/feature-branch or PWD as target
    1. [x] Provide own library to wrap functions with (router, serving, ...)
-   1. [ ] Think about monorepos in Deno (and how to organize them)
+   1. [x] Think about monorepos in Deno (and how to organize them)
    1. [ ] Testing
-   1. [ ] Script running (`./cache.sh`, ...)
+   1. [x] Script running (`./cache.sh`, ...)
