--- cache.sh
@@ -1,2 +1,0 @@
-deno cache --lock=lock.json --lock-write deps.ts
-deno run --allow-read --allow-write scripts/sortLock.ts
