--- execution.sh
@@ -1,1 +1,0 @@
-deno run --unstable --allow-write --allow-net --allow-read --allow-run src/execution/index.ts
