+++ lock.json
@@ -62,6 +62,17 @@
   "https://deno.land/std@0.59.0/uuid/v1.ts": "0e37292975b956f6591c057398cfb4fbd9a06f22b45e22224fe9a5a9373126d7",
   "https://deno.land/std@0.59.0/uuid/v4.ts": "ddfce12c2e271a9422751df420d755f44197856f41c6cd37e2606e168d978d58",
   "https://deno.land/std@0.59.0/uuid/v5.ts": "952fd2f65b5dcc272ae7ae63176799ada113e6cb85a43975888cfd05639526ca",
+  "https://deno.land/std@0.60.0/_util/assert.ts": "e1f76e77c5ccb5a8e0dbbbe6cce3a56d2556c8cb5a9a8802fc9565af72462149",
+  "https://deno.land/std@0.60.0/path/_constants.ts": "e11f32a36644e04dce243f3c1f30c02cc5d149827f48a755628176f0707cfc70",
+  "https://deno.land/std@0.60.0/path/_globrex.ts": "696996deed47bd98fcab9de0034233380ead443a5cd702631151c6a71e3d21a2",
+  "https://deno.land/std@0.60.0/path/_interface.ts": "5876f91d35fd42624893a4aaddaee352144e1f46e719f1dde6511bab7c3c1029",
+  "https://deno.land/std@0.60.0/path/_util.ts": "feacfbe104df3baf2a08ca80263d6a9cf63f9c32e7c9cbe54602ec12864a72e9",
+  "https://deno.land/std@0.60.0/path/common.ts": "e4ec66a7416d56f60331b66e27a8a4f08c7b1cf48e350271cb69754a01cf5c04",
+  "https://deno.land/std@0.60.0/path/glob.ts": "ff5c4d04d643083dc5c8a46da83b8c836934155c2a7ad72675c43092b2a435ea",
+  "https://deno.land/std@0.60.0/path/mod.ts": "6de8885c2534757097818e302becd1cefcbc4c28ac022cc279e612ee04e8cfd1",
+  "https://deno.land/std@0.60.0/path/posix.ts": "514c39f5f7ca45bee2ace2a2e83da74dde5ff08fba837c71f86f459abe846fd7",
+  "https://deno.land/std@0.60.0/path/separator.ts": "9dd15d46ff84a16e13554f56af7fee1f85f8d0f379efbbe60ac066a60561f036",
+  "https://deno.land/std@0.60.0/path/win32.ts": "91dfd12ddaabec7fa4a8052e16ec56a260722a505e3bb8dfc35041651af38322",
   "https://deno.land/x/accepts@1.0.0/deps.ts": "230c8e4888c7b745a2b6c33cedc53c25f77fadb089c0dddba978418909497aea",
   "https://deno.land/x/accepts@1.0.0/mod.ts": "928576bb8b922e900fc629f80050a13248cd3bae0ee05d70e0dc69243e71fe15",
   "https://deno.land/x/content_type/mod.ts": "5dc435e84b665c2968d6845c573569ddeef65a41baa9a04c51e1c5e927806583",
@@ -116,7 +127,7 @@
   "https://deno.land/x/isIP@1.0.0/mod.ts": "bd55c2180f7275930d9b1ef0b8f031108625ae3f9b87a1fa75cc8ee15d485ff8",
   "https://deno.land/x/media_typer/mod.ts": "01aebf79f0ecf43897a663dca9730d2dbd2a698eac9d297a6ac26b4b647bb7cc",
   "https://deno.land/x/media_types/db.ts": "56a9deab286b6283e1df021d74ee3319353f27f7827716b6443427fff2fc6e24",
-  "https://deno.land/x/media_types/deps.ts": "753cb10049c828a5f9d0253b51b885824f048499aad9ff33461a9bd23d986b1c",
+  "https://deno.land/x/media_types/deps.ts": "c0f528f906ad64d2c59f87e82c8cb1a972ed5e3a814b2eda959742a914d117c3",
   "https://deno.land/x/media_types/mod.ts": "94141d7c415fcdad350fec9d36d77c18334efe25766db6f37d34709c896881ed",
   "https://deno.land/x/media_types@v2.3.7/db.ts": "56a9deab286b6283e1df021d74ee3319353f27f7827716b6443427fff2fc6e24",
   "https://deno.land/x/media_types@v2.3.7/deps.ts": "0ae24f129e2075dc18adeba5574319ec416a6053f2b8a7772d562deda5564bcc",
