+++ packages/execution/deps.ts
@@ -1,0 +1,3 @@
+export * from "../../deps.ts";
+export * from "../registry/mod.ts";
+export * from "../shared/mod.ts";
