+++ packages/registry/deps.ts
@@ -1,0 +1,2 @@
+export * from "../../deps.ts";
+export * from "../shared/mod.ts";
