+++ packages/registry/mod.ts
@@ -1,0 +1,1 @@
+export * from "./src/registry.ts";
