+++ packages/scripts/cache.sh
@@ -1,0 +1,2 @@
+deno cache --lock=lock.json --lock-write deps.ts
+deno run --allow-read=$PWD --allow-write=$PWD packages/scripts/sortLock.ts
