+++ packages/shared/src/dirExists.ts
