+++ packages/shared/src/path.ts
