+++ packages/shared/src/pingServer.ts
