+++ packages/templates/bundle.ts
@@ -1,0 +1,6 @@
+// TODO: reference to github/jsdeliver
+import { serve } from "./packages/micro/mod.ts";
+
+import fn from "./registry_in/%SCRIPT_NAME%/index.ts";
+
+await serve(fn, %PORT%);
