--- templates/bundle.ts
@@ -1,5 +1,0 @@
-import { serve } from "./src/micro/mod.ts";
-
-import fn from "./registry_in/%SCRIPT_NAME%/index.ts";
-
-await serve(fn, %PORT%);
