--- todos.md
@@ -1,22 +1,0 @@
-# SECURITY TODOS
-
-- [x] Put registry back to linked volume -> you want to mount to external volume on cloud (Digitalocean f.e.)
-  - [x] put this path in a variable loaded by shell script
-- [x] consolidate shell scripts into 1
-  - [x] ingest volume path to registry
-  - [x] MAYBE remove + create docker volume to registry `docker volume create ...`
-  - [x] docker-compose up -d
-- [x] docker-compose
-  - [x] add registry volume
-- [x] registry
-  - [x] move away from potentially allowing workers all deno permissions and refactor it to a network whitelist
-- [x] execution
-  - [x] ingest permissions for network whitelist and pass to orchestration
-- [x] container security
-  - [x] lock down allow read/write?
-- [ ] extract orchestration layer to separate module
-  - [ ] try toAsyncIterator for spawning deno processes
-  - [ ] spawning and killing shouldn't exist "inside" execution layer
-  - [ ] potentially extract into separate dockerfile as well
-- [ ] move to monorepo and ingest deps per package (and ingest own deps from jsdeliver/deno.land/...)
-  - [ ] maybe run publish script that rewrites deps per package
